+++ project/date_dimension/tests/conftest.py
@@ -1,0 +1,1 @@
+from date_dimension.tests.factories import *