+++ project/elektron/settings/base.py
@@ -67,6 +67,7 @@ INSTALLED_APPS = [
     'users',
     'pages',
     'media_items',
+    'date_dimension',
 ]
 
 MIDDLEWARE = [
@@ -88,6 +89,7 @@ TEMPLATES = [
         'DIRS': [
             os.path.join(BASE_DIR, 'pages/templates'),
             os.path.join(BASE_DIR, 'users/templates'),
+            os.path.join(BASE_DIR, 'media_items/templates'),
             os.path.join(BASE_DIR, 'templates'),
         ],
         'APP_DIRS': True,