+++ project/media_items/tests/test_upload_processing_integration.py
@@ -19,7 +19,7 @@ def test_sign_upload_request_success(authenticated_client, img, env):
     with(env['remote_path']):
         c, u = authenticated_client
 
-        image_key = request_upload(c, img)
+        image_key = request_upload(c, u, img)
         media_id = parse_id_from_key(u.id, image_key)
 
         remote_file = mock_upload(img['local_path'], env['remote_path'], image_key)
@@ -33,9 +33,9 @@ def test_sign_upload_request_success(authenticated_client, img, env):
 def parse_id_from_key(user_id, key):
     m = match_image_key(str(user_id), key)
     assert m is not None
-    id = m.group('image_id')
-    assert id is not None
-    return id
+    image_id = m.group('image_id')
+    assert image_id is not None
+    return image_id
 
 
 def assert_processing(e, a):
@@ -75,8 +75,9 @@ def to_date(s):
     return dt
 
 
-def request_upload(client, img):
-    response = client.post('/media/request-upload/', {'mime_type': img['mime_type']})
+def request_upload(client, user, img):
+    request_url = '/media/%s/request-upload/' % user.id
+    response = client.post(request_url, {'mime_type': img['mime_type']})
     assert response.status_code == 201
     path = response['Location']
     return urlparse(path).path
@@ -98,7 +99,7 @@ def invoke_processor(image_key):
 def build_clean_processor():
     cwd = processor_project_dir()
     cmd = ['./gradlew', '--quiet', 'clean', 'fatJar']
-    exec(cwd, cmd)
+    exec_cmd(cwd, cmd)
 
 
 def run_processor(path):
@@ -115,17 +116,17 @@ def run_processor(path):
         '-f',
         path
     ]
-    exec(cwd, cmd)
+    exec_cmd(cwd, cmd)
 
 
-def exec(cwd, cmd):
+def exec_cmd(cwd, cmd):
     subprocess.run(
         args=cmd,
         cwd=cwd,
-        #stdout=subprocess.DEVNULL,
-        #stderr=subprocess.DEVNULL
+        # stdout=subprocess.DEVNULL,
+        # stderr=subprocess.DEVNULL
     )
 
 
 def processor_project_dir():
-    return os.path.realpath( '%s/../processor' % os.getcwd() )
+    return os.path.realpath('%s/../processor' % os.getcwd())