+++ CHANGES.md
@@ -5,6 +5,7 @@ Elektron Change Log
 -----------------
 
 - [#36] Eliminate requirement to include create date when requesting image upload.
+- [#37] Views and model for date-based views of media (collections, albums, items).
 
 
 0.26 (2019-04-30)